added checkpoint system
You now set the directory name for files in the SETTINGS.INI.  If you try to generate a file that already exists in that directory it will skip it. This helps you basically checkpoint progress if you have to stop and pick up later. (Although ChatGPT may generate a new list of actors).